moved to useDataCalculations
